sys/net/application_layer/sock_dns_mock: add module for mocking sock_dns
