+++ sys/net/application_layer/sock_dns_mock/Makefile
@@ -1,0 +1,1 @@
+include $(RIOTBASE)/Makefile.base
