+++ tests/netutils/Makefile
@@ -12,7 +12,6 @@ USEMODULE += gnrc_ipv6
 USEMODULE += ipv4_addr
 USEMODULE += ipv6_addr
 
-# pretend to include sock_dns
-CFLAGS += -DMODULE_SOCK_DNS=1
+USEMODULE += sock_dns_mock
 
 include $(RIOTBASE)/Makefile.include
